# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /fav-stocks:
    get:
      tags:
        - favourite-stocks-controller
      description: Get favorite stocks for particular stockUser.
      operationId: getFavouriteStocks
      responses:
        "403":
          description: "For User: If stockUser wants to see others favourite stocks"
        "200":
          description: "For Admin and User: Admin can access anyone. User can access\
            \ only his fav stocks."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockValue'
      security:
        - Bearer Authentication: []
    post:
      tags:
        - favourite-stocks-controller
      description: Add stocks to favourite.
      operationId: addStockToFavourite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteStockManipulationDTO'
        required: true
      responses:
        "201":
          description: "For Admin and User: Successfully added stock to the list of\
            \ fav. "
        "403":
          description: "For User: If user wants to add fav stock for other user"
        "404":
          description: "For Admin: If user was not found in the database. Or also\
            \ For Admin and User: If stock with provided name was not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      security:
        - Bearer Authentication: []
    delete:
      tags:
        - favourite-stocks-controller
      description: Delete stock from the list of favourites.
      operationId: deleteFromFavourite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteStockManipulationDTO'
        required: true
      responses:
        "403":
          description: "For User: If user wants to delete fav stock for other user"
        "204":
          description: "For Admin and User: Successfully deleted stock from the list\
            \ of fav. "
        "404":
          description: "For Admin: If user was not found in the database. Or also\
            \ For Admin and User: If stock with provided name was not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      security:
        - Bearer Authentication: []
...
components:
  schemas:
...
    ErrorModel:
      type: object
      properties:
        error:
          type: string
...
    FavouriteStockManipulationDTO:
      type: object
      properties:
        symbol:
          type: string
        userId:
          type: integer
          format: int32
...
    StockValue:
      type: object
      properties:
        datetime:
          type: string
        open:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        close:
          type: number
          format: double
        volume:
          type: integer
          format: int32
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT